Refactor FML command line out of main.rs
