+++ components/support/nimbus-fml/src/lib.rs
@@ -2,6 +2,8 @@
 * License, v. 2.0. If a copy of the MPL was not distributed with this
 * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
+mod backends;
+pub mod command_line;
 pub(crate) mod defaults_merger;
 pub mod error;
 pub(crate) mod frontend;
@@ -17,7 +19,6 @@ cfg_if::cfg_if! {
 }
 
 #[cfg(test)]
-#[allow(dead_code)]
 pub mod fixtures;
 
 const SUPPORT_URL_LOADING: bool = true;
